assignment-03/control-unit-subsystem/requirements.txt
pyserial==3.5
paho-mqtt==1.6.1
flask==3.0.0
flask_cors==4.0.0